+++ CHANGELOG.md
@@ -9,6 +9,9 @@
     * Qwen3.5-2B-Q4_K_M
     * Qwen3.5-0.8B-Q8_0
     * LFM2.5-350M-Q8_0
+* LocalLLM
+    * put __call__ in try-catch block and return "" in case of an error (e.g. context to big for the model)
+    * in __call__ trace more arguments if requested
 
 ## 0.2.0