+++ CHANGELOG.md
@@ -9,6 +9,8 @@
     * Qwen3.5-2B-Q4_K_M
     * Qwen3.5-0.8B-Q8_0
     * LFM2.5-350M-Q8_0
+    * gemma-4-e2b-it-Q8_0
+    * gemma-4-E2B-it-Q4_K_M
 * LocalLLM
     * put __call__ in try-catch block and return [None] in case of an error (e.g. context to big for the model) + log the error in case of trace
     * in __call__ trace more arguments if requested